requirements.txt
flask
requests
python-dotenv
azure-cognitiveservices-vision-computervision
msrest
